ordering app along with bill done

+++ menu/models.py
@@ -9,6 +9,6 @@ class RestaurantMenu(models.Model):
     foodtype = models.CharField(max_length=10, choices=FoodType.choices)
     active = models.BooleanField(default=True)
     name = models.CharField(max_length=100, unique=True)
-    price = models.DecimalField(decimal_places=2, max_digits=10, default=0)
+    price = models.IntegerField(default=0)
